// sourcecode/BTrak.Packages/snova-project-management/PM-module-components/pm-components/src/lib/snovasys-projects/models/project.ts
export class Project {
  projectResponsiblePersonId: string;
  projectId: string;
  projectTypeId: string;
  isArchived: boolean;
  projectName: string;
  projectTypeName: string;
  projectStatusColor: string;
  userName: string;
  createdDateTime: Date;
  fullName: string;
  roleName: string;
  profileImage: string;
  projectResponsiblePerson:any;
  activeGoalCount: number;
  numberOfReds: number;
  timeStamp: any;
  projectLabel: string;
  testSuiteCount: number;
  testRunCount: number;
  milestoneCount: number;
  totalCount: number;
  reportCount: number;
  casesCount: number;
  auditsCount: number;
  conductsCount: number;
  auditActionsCount: number;
  auditReportsCount: number;
  auditQuestionsCount: number;
  isDateTimeConfiguration: boolean;
  isSprintsConfiguration: boolean;
  viewGoalsPermissionCount: number;
  projectStartDate: Date;
  projectEndDate: Date;
  timeZoneOffSet: any;
}